[Update] 토론 목록 검색 기능 일반 조회와 통합
