+++ components/debates/debates/DebatesList.tsx
@@ -1,9 +1,9 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useInView } from "react-intersection-observer";
 
 import { checkCategory } from "../../../utils/debates/checkCategory";
 import { useGetDebates } from "../../../utils/queries/debates";
-import { useSelect } from "../../../utils/common/useInputSelect";
+import { useInput, useSelect } from "../../../utils/common/useInputSelect";
 import { COMMENT_ORDER } from "../../../utils/common/constant";
 
 import { DebateOfDebates } from "../../../types";
@@ -11,8 +11,10 @@ import { DebateOfDebates } from "../../../types";
 export default function DebatesList({ list }: { list: string[] }) {
   const { ref, inView } = useInView();
   const orderSelect = useSelect(COMMENT_ORDER[0][1], refetch);
+  const [isSearchListOn, setIsSearchListOn] = useState<boolean>(false);
+  const search = useInput("", "");
 
-  const debates = useGetDebates(orderSelect.value);
+  const debates = useGetDebates(search.value, orderSelect.value);
 
   useEffect(() => {
     if (inView && debates.hasNextPage) debates.fetchNextPage();
@@ -36,6 +38,33 @@ export default function DebatesList({ list }: { list: string[] }) {
 
   return (
     <div>
+      {isSearchListOn ? (
+        <div>
+          <input {...search.attribute} disabled />
+          <button
+            onClick={() => {
+              search.setValue("");
+              refetch();
+              setIsSearchListOn(false);
+            }}
+          >
+            X
+          </button>
+        </div>
+      ) : (
+        <div>
+          <input {...search.attribute} />
+          <button
+            onClick={() => {
+              refetch();
+              setIsSearchListOn(true);
+            }}
+            disabled={!search.value}
+          >
+            검색
+          </button>
+        </div>
+      )}
       <select {...orderSelect.attribute}>
         {COMMENT_ORDER.map((order) => (
           <option key={order[0]} value={order[1]}>
