+++ utils/queries/debates.ts
@@ -16,7 +16,6 @@ import {
   getDebate,
   getDebates,
   getDebatesHeart,
-  getDebatesSearch,
   patchDebate,
   postDebate,
 } from "../../api/debates";
@@ -24,10 +23,10 @@ import { queryStr } from ".";
 
 import { Debate, DebatePost, DebatePatch, User } from "../../types";
 
-export const useGetDebates = (order: string) => {
+export const useGetDebates = (searchValue: string, order: string) => {
   const query = useInfiniteQuery(
     [queryStr.debates],
-    ({ pageParam = 0 }) => getDebates(pageParam, order),
+    ({ pageParam = 0 }) => getDebates(searchValue, pageParam, order),
     {
       getNextPageParam: (lastPage) =>
         !lastPage.isLast ? lastPage.nextPage : undefined,
@@ -40,18 +39,6 @@ export const useGetDebatesHeart = (userId: string, order: string) => {
   const query = useInfiniteQuery(
     [queryStr.debates, "heart"],
     ({ pageParam = 0 }) => getDebatesHeart(userId, pageParam, order),
-    {
-      getNextPageParam: (lastPage) =>
-        !lastPage.isLast ? lastPage.nextPage : undefined,
-    },
-  );
-  return query;
-};
-
-export const useGetDebatesSearch = (searchValue: string, order: string) => {
-  const query = useInfiniteQuery(
-    [queryStr.debates, "search"],
-    ({ pageParam = 0 }) => getDebatesSearch(searchValue, pageParam, order),
     {
       getNextPageParam: (lastPage) =>
         !lastPage.isLast ? lastPage.nextPage : undefined,
